move testutil/known_values.go to testutil/testdata
