+++ testutil/testdata/known_values.go
@@ -1,4 +1,4 @@
-package testutil
+package testdata
 
 const (
 	// Tests expect a ledger device initialized to the following mnemonic
